requirements.txt
# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# API Server
flask>=2.3.0
flask-cors>=4.0.0
redis>=5.0.0

